% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{C}
\alias{C}
\title{Numerical concatenation}
\usage{
C(...)
}
\arguments{
\item{...}{Numerics}
}
\value{
The digits from input concatenated. Returned as numeric.
}
\description{
Numerical concatenation
}
\examples{

RG::C(2,3,9)
C(1,5,1)
}
