% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{Load}
\alias{Load}
\title{Load this package}
\usage{
Load()
}
\value{
Nothing!
}
\description{
This function is called within every other function of this package, and calls \code{library(RG)} if it isn't attached yet. This ensures that you only need to use the \code{RG::} prefix once.
}
\examples{
{
RG::Load()
C(2,3,9)}
}
