% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendStrings.R
\name{+}
\alias{+}
\title{Addition}
\usage{
"+"(e1, e2)
}
\arguments{
\item{e1}{LHS}

\item{e2}{RHS}
}
\value{
A string or numeric.
}
\description{
Modified addition operator to work with strings.
}
\details{
If any of the arguments is of class \code{character}, the two arguments are concatenated and returned as string. Else the base implementation of \code{+} is called.
}
\examples{
{
1+1
"a"+"b"
"a"+1+2
1+2+"a"
}
}
