% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg.R
\name{rg}
\alias{rg}
\title{This package is to simulate dataset following linear regression model}
\source{
The data comes from simulation.
}
\usage{
rg(n, nv)
}
\arguments{
\item{n}{the required sample size}

\item{nv}{the required number of variables}
}
\value{
The generated dataframe with simple size=n and nv number of independent variables
}
\description{
Choose the n sample size and nv number of variables to get the generated dataset
}
\details{
This is package built by Fadhaa Ali
}
\references{
Nagarajan R, Scutari M, Lebre S (2013). "Bayesian Networks in R with Applications in Systems Biology". Springer.
}
\author{
Author: Fadhaa Ali
}
\concept{Linear Regression.}
