% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitStrings.R
\name{s}
\alias{s}
\title{Split a vector into single characters or digits}
\usage{
s(x, on = "")
}
\arguments{
\item{x}{Vector to split. Can be numeric or character.}

\item{on}{Character string to split on, defaults to \code{""}.}
}
\value{
If \code{length(x) == 1}, returns split as a vector, else as a list.
}
\description{
Takes a string and returns all separated characters. If the input is numeric, the input is split into its digits and returned as numerics.
}
\examples{
s("bla")
s(c("bla", "bla"))
s(155)
s(c(151, 239))
}
