% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringRepetition.R
\name{*}
\alias{*}
\title{Character repetition}
\usage{
"*"(e1, e2)
}
\arguments{
\item{e1}{}

\item{e2}{}
}
\value{
Character if \code{e1} is character, regular output otherwise.
}
\description{
Shorthand for \code{paste(rep(e1, each = e2), collapse = "")}. Works like regular multiplication if \code{e1} is not character.
}
\examples{
2*2
"A"*2
"A"*2+1
"A"*(2+1)
}
